+++ tests/testthat/funmatch/R/test.R
@@ -1,0 +1,3 @@
+mean <- 10
+
+pi <- function() 3